Split out ImageResourceEntryNode, and add support for cursor resources.
