+++ ILSpy/TreeNodes/ResourceNodes/ResourceEntryNode.cs
@@ -91,61 +91,4 @@ namespace ICSharpCode.ILSpy.TreeNodes
 			return true;
 		}
 	}
-
-	[Export(typeof(IResourceNodeFactory))]
-	sealed class ImageResourceNodeFactory : IResourceNodeFactory
-	{
-		static readonly string[] imageFileExtensions = { ".png", ".gif", ".bmp", ".jpg", ".ico" };
-
-		public ILSpyTreeNode CreateNode(Mono.Cecil.Resource resource)
-		{
-			EmbeddedResource er = resource as EmbeddedResource;
-			if (er != null) {
-				return CreateNode(er.Name, er.GetResourceStream());
-			}
-			return null;
-		}
-
-		public ILSpyTreeNode CreateNode(string key, Stream data)
-		{
-			foreach (string fileExt in imageFileExtensions) {
-				if (key.EndsWith(fileExt, StringComparison.OrdinalIgnoreCase))
-					return new ImageResourceEntryNode(key, data);
-			}
-			return null;
-		}
-	}
-
-	sealed class ImageResourceEntryNode : ResourceEntryNode
-	{
-		public ImageResourceEntryNode(string key, Stream data)
-			: base(key, data)
-		{
-		}
-
-		public override object Icon
-		{
-			get { return Images.ResourceImage; }
-		}
-
-		internal override bool View(DecompilerTextView textView)
-		{
-			try {
-				AvalonEditTextOutput output = new AvalonEditTextOutput();
-				Data.Position = 0;
-				BitmapImage image = new BitmapImage();
-				image.BeginInit();
-				image.StreamSource = Data;
-				image.EndInit();
-				output.AddUIElement(() => new Image { Source = image });
-				output.WriteLine();
-				output.AddButton(Images.Save, "Save", delegate { Save(null); });
-				textView.ShowNode(output, this, null);
-				return true;
-			}
-			catch (Exception) {
-				return false;
-			}
-		}
-	}
 }
